python/requirements.txt
fastapi==0.84.0
uvicorn==0.11.5
mysql-connector-python
pandas
sqlalchemy 